move common publickey test util to its own file
