+++ clientapi/auth/login_publickey_test.go
@@ -20,116 +20,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/matrix-org/dendrite/setup/config"
-	uapi "github.com/matrix-org/dendrite/userapi/api"
 	"github.com/stretchr/testify/assert"
 )
-
-type fakePublicKeyUserApi struct {
-	UserInternalAPIForLogin
-	uapi.UserLoginAPI
-	uapi.ClientUserAPI
-	DeletedTokens []string
-}
-
-func (ua *fakePublicKeyUserApi) QueryAccountAvailability(ctx context.Context, req *uapi.QueryAccountAvailabilityRequest, res *uapi.QueryAccountAvailabilityResponse) error {
-	if req.Localpart == "does_not_exist" {
-		res.Available = true
-		return nil
-	}
-
-	res.Available = false
-	return nil
-}
-
-func (ua *fakePublicKeyUserApi) QueryAccountByPassword(ctx context.Context, req *uapi.QueryAccountByPasswordRequest, res *uapi.QueryAccountByPasswordResponse) error {
-	if req.PlaintextPassword == "invalidpassword" {
-		res.Account = nil
-		return nil
-	}
-	res.Exists = true
-	res.Account = &uapi.Account{}
-	return nil
-}
-
-func (ua *fakePublicKeyUserApi) PerformLoginTokenDeletion(ctx context.Context, req *uapi.PerformLoginTokenDeletionRequest, res *uapi.PerformLoginTokenDeletionResponse) error {
-	ua.DeletedTokens = append(ua.DeletedTokens, req.Token)
-	return nil
-}
-
-func (ua *fakePublicKeyUserApi) PerformLoginTokenCreation(ctx context.Context, req *uapi.PerformLoginTokenCreationRequest, res *uapi.PerformLoginTokenCreationResponse) error {
-	return nil
-}
-
-func (*fakePublicKeyUserApi) QueryLoginToken(ctx context.Context, req *uapi.QueryLoginTokenRequest, res *uapi.QueryLoginTokenResponse) error {
-	if req.Token == "invalidtoken" {
-		return nil
-	}
-
-	res.Data = &uapi.LoginTokenData{UserID: "@auser:example.com"}
-	return nil
-}
-
-func initializeTestUserInteractive() *UserInteractive {
-	userInteractive := UserInteractive{
-		Flows:    []userInteractiveFlow{},
-		Types:    make(map[string]Type),
-		Sessions: make(map[string][]string),
-		Params:   make(map[string]interface{}),
-	}
-
-	return &userInteractive
-}
-
-func initializeTestConfig() *config.ClientAPI {
-	chainIds := []int{4}
-	cfg := &config.ClientAPI{
-		Matrix: &config.Global{
-			ServerName: serverName,
-		},
-		PublicKeyAuthentication: config.PublicKeyAuthentication{
-			Ethereum: config.EthereumAuthConfig{
-				Enabled:  true,
-				Version:  1,
-				ChainIDs: chainIds,
-			},
-		},
-	}
-
-	return cfg
-}
-
-func testPublicKeySession(
-	ctx *context.Context,
-	cfg *config.ClientAPI,
-	userInteractive *UserInteractive,
-	userAPI *fakePublicKeyUserApi,
-
-) string {
-	emptyAuth := struct {
-		Body string
-	}{
-		Body: `{
-			"type": "m.login.publickey"
-		 }`,
-	}
-
-	_, cleanup, err := LoginFromJSONReader(
-		*ctx,
-		strings.NewReader(emptyAuth.Body),
-		userAPI,
-		userAPI,
-		userAPI,
-		userInteractive,
-		cfg)
-
-	if cleanup != nil {
-		cleanup(*ctx, nil)
-	}
-
-	json := err.JSON.(Challenge)
-	return json.Session
-}
 
 func TestLoginPublicKeyNewSession(t *testing.T) {
 	// Setup
